feat: enable expressions for primitive values on Props form (#2049)
* refactor(libs-frontend): use react context instead of property drilling (#1989)

Define context which will contain all the data to render and validate form. In future commits
there will be even more properties added and they will be used in bigger number of places,
so using context is more prefereable than passing parameters all over the components tree

* feat(libs-frontend): add support of expressions for number and boolean (#1989)

Add new component which is capable to switch between primiteive field and CodeMirror. Also implement
ability to save expression for boolean and number fields on Props form

* feat(libs-frontend): validate expressions of Props form (#1989)

evaluate and validate expression result when Props form is submited

* fix(libs-frontend-view): fix infinite recursion (#1989)

createAutocompleOptions function accepts and object with application state. When app contains an
input component (it will be present in state) - this caused an infinite recursion call which throws
error about max stack size reached. Some properties on an input causes this problem. Added
additional validation: if the object in state is a DOM node- do not process it recursively

* style(libs-frontend-view): fix styling of CodeMirror field (#1989)

fix styling of CodeMirror field to be the same as other fields. This is requeired so that there is
no much difference between basic ant design components (inputs, selects) and CodeMirror when the
toggle button is clicked and one component is replaced with another

* feat(libs-frontend-view): add autocompletion for expressions field (#1989)

* refactor(apps): cleanup console.log statements

remove console.log statements which were commented or were used for debugging purposes. They made
reading console logs in local and CI environment harder

Co-authored-by: Vladyslav Polishchuk <vladyslav@Vladyslavs-MacBook-Pro.local>