fix(apps-cli): fix key/value seed data for enums (#2056)
When on the Props form some value is selected from the dropdown (EnumType) - pascal case value is
used and passed as a parameter to the component. Such values are invalid and they are ignored by
components. So most of the Enum properties do not work for many of the atoms. For example there is
"shape" dropdown for Button atom exists. The corresponding dropdown had values equal to "Square" and
"Circle", but this are invalid values, "square" and "circle" must be used instead. So this commit
swaps key and value, and now all dropdowns will render on UI user-friendly Pascal case name and will
use lower-case strings as values.