feat: improve props UI (#2065)
* refactor(libs-frontend-view): refactor form validation

After refactoring in commit 1fc1a0124edbd3ba1792f2a6e8346c1f86236dd1 - new method is available to
deeply evaluate expressions in props. Use this method instead of manual implementation

* feat(libs-frontend): implement expressions for enums

Allow to set expression on Props form for fields of type Enum

* feat(libs-frontend): implement expressions for arrays

Allow to set expression on Props form for fields of type Array.
Fix script error inside "mapDeep" method when value is nullish.

* style(libs-frontend-view): fix height of CodeMirror field

Now height of CodeMirror field is going to be auto (one line) by default. If some value is multiline
- the field will auto-adjust until height of 150px is reached. Starting from that point inner scroll
appears. This should make most of the forms much more compact, since strings props are mostly
oneliners

* feat(libs-frontend): add support of expressions for ActionType fields

Now it is possible to either select action in dropdown list or toggle to CodeMirror field and write
a function there

* fix(libs-frontend): fix always empty list of Elements in dropdown

On props form the list of available elements was always empty. Now it is populated with valid
Elements form elementsTree

* feat(libs-frontend): add expression support for RenderProps and ReactNode types

Now it is possible to specify a function which returns JSX for RenderProps and ReactNode fields on
Props form